feat: ability to fetch router from huma context
This allows us to query the router for information.  I've added a
GetOperation method so we can query for information about a given
operation.  This is useful in hypermedia APIs because it allows
us to link across different endpoints without having to hardwire
paths.  Instead, the cross-referencing is done via operation id.